pyproject.toml
[project]
name = "openacm"
version = "0.1.0"
description = "OpenACM - Multi-channel AI Assistant with system control"
requires-python = ">=3.12"

dependencies = [
    # Core LLM
    "litellm>=1.50",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    
    # Web
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "websockets>=13.0",
    
    # Channels
    "discord.py>=2.4",
    "python-telegram-bot>=21.0",
    "httpx>=0.27",
    
    # Storage
    "aiosqlite>=0.20",
    
    # Tools
    "duckduckgo-search>=6.0",
    "psutil>=6.0",
    
    # Google Services
    "google-api-python-client>=2.130",
    "google-auth-oauthlib>=1.2",
    "google-auth-httplib2>=0.2",
    
    # Utils
    "rich>=13.0",
    "structlog>=24.0",
    "python-dotenv>=1.0",
    
    # Multimodal & Crypto
    "mss>=9.0",
    "Pillow>=10.0",
    "cryptography>=41.0",
    
    # Tier-1 Agentic powers
    "chromadb>=0.5",
    "sentence-transformers>=3.3",
    "playwright>=1.49",
    "jupyter-client>=8.6",
    "ipykernel>=6.29"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[project.scripts]
openacm = "openacm.__main__:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
